proto: Add agent grpc support
