+++ go.mod
@@ -14,5 +14,6 @@ require (
 	github.com/stretchr/testify v1.7.0
 	golang.org/x/crypto v0.0.0-20201016220609-9e8e0b390897 // indirect
 	golang.org/x/sys v0.0.0-20201101102859-da207088b7d1 // indirect
+	google.golang.org/grpc v1.32.0
 	google.golang.org/protobuf v1.25.0
 )
