+++ proto/doc.go
@@ -1,4 +1,4 @@
 package proto
 
 //go:generate protoc --go_out=. --go_opt=paths=source_relative task.proto
-//go:generate protoc --go_out=. --go_opt=paths=source_relative agent.proto
+//go:generate protoc --go_out=. --go_opt=paths=source_relative --go-grpc_out=. --go-grpc_opt=paths=source_relative agent.proto
